Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
numpy==1.26.4
opencv-python-headless==4.10.0.84
pydantic==2.9.2
google-genai>=1.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
